frontend/requirements.txt
streamlit==1.31.0
requests==2.31.0
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
altair==5.2.0
